story routes finished and ready to be reviewed

+++ app/api/story_routes.py
@@ -1,30 +1,25 @@
-from flask import Blueprint, request #when do I use request and when do I use a form????
+from flask import Blueprint, jsonify, request
 from flask_login import login_required, current_user
 from app.forms import StoryEditorForm
 from app.models import db, Story, User, Comment, Like
+from sqlalchemy.orm import joinedload
+from datetime import datetime
 
 story_routes = Blueprint('stories', __name__)
 
-
-#MAYBE DONE - EXCEPT VALIDATION ERRORS
+#DONE
 # get all stories
 @story_routes.route('/stories')
 def get_all_stories():
-    stories = Story.query.(Story)options(joinedload(Story.user)).order_by(Story.created_date.desc())
-    print(stories)
-    if not stories:
-        raise Exception('there are no stories')
-    obj_stories = [story.author_info_to_dict() for story in stories]
-    return {'Stories': obj_stories}
+    stories = Story.query.order_by(Story.created_date.desc()).all()
+    if not bool(stories): 
+        return jsonify({'message': 'Stories could not be found'}), 404
+    return jsonify({'Stories': [story.preview_story_to_dict() for story in stories]}), 200
 
-
+#DONE
+# create a story
 @story_routes.route('/stories', methods=['POST'])
 @login_required
-# def get_stories_feed():
-#     stories = Story.query().options(joinedload(Story.user)).filter(Story.user.in_(current_user.follows)).order_by(Story.created_date.desc())
-#     obj_stories = [story.author_info_to_dict() for story in stories]
-#     return {'Stories': obj_stories}
-
 def create_story():
     form = StoryEditorForm()
     form['csrf_token'].data = request.cookies['csrf_token']
@@ -38,40 +33,55 @@ def create_story():
         db.session.add(story)
         db.session.commit()
         return story.to_dict()
+    else: 
+        return jsonify({'message': 'Story needs the have a title and content'}), 400
    
 
-#NOT DONE
+#DONE
 # get a single story
-@story_routes.route('/stories/<int:story_id>', methods=['GET', 'PUT', 'DELETE'])
-@login_required
+@story_routes.route('/stories/<int:story_id>')
 def get_one_story(story_id):
-    story = Story.query.(story_id).options(joinedload(Story.user))
-    number_of_likes = story.num_comments()
-    number_of_comments = story.num_likes()
-    story_obj = story.to_dict()
-    story_obj['num_likes'] = number_of_likes
-    story_obj['num_comments'] = number_of_comments
-    print(story)
-    return story_obj
-
-def update_one_story(story_id):
-    form = StoryEditorForm()
-    form['csrf_token'].data = request.cookies['csrf_token']
-    story = Story.query.get(story_id)
-    if story.user_id == current_user.id:
-        if form.validate_on_submit():
-            story.title = form.data['title'],
-            story.content=form.data['content'],
-            story.image_url=form.data['image_url']
-        db.session.add(story)
-        db.session.commit()
-        return story.to_dict()
-    
-def delete_one_story(story_id):
-    story = Story.query.get(story_id)
+    story = Story.query.filter(Story.id == story_id).first()
     if story:
-        if story.user_id == current_user.id:
-            db.session.delete(story)
-            db.session.commit()
-            return story.to_dict()
-        
+        story_obj = story.full_story_to_dict()
+        return jsonify(story_obj), 200
+    else:
+        return jsonify({'message': 'Story could not be found'}), 404
+
+#DELETE IS DONE, EDIT IS DONE
+# edit & delete a single story 
+@story_routes.route('/stories/<int:story_id>', methods=['PUT', 'DELETE'])
+@login_required
+def update_one_story(story_id):
+    if request.method == 'PUT':
+        form = StoryEditorForm()
+        form['csrf_token'].data = request.cookies['csrf_token']
+        story = Story.query.filter(Story.id == story_id).first()
+        if story:
+            if story.user_id == current_user.id:
+                if form.validate_on_submit():
+                    story.title = form.data['title'] or story.title
+                    story.content=form.data['content'] or story.content
+                    story.image_url=form.data['image_url'] or story.image_url
+                    story.updated_date = datetime.now()
+                    db.session.add(story)
+                    db.session.commit()
+                    return jsonify(story.to_dict()), 200
+                else: 
+                    return jsonify({'message': 'Story needs the have a title and content'}), 400
+            else: 
+                return jsonify({'message': 'Users can only edit their own stories'}), 403
+        else:
+            return jsonify({'message': 'Story could not be found'}), 404
+
+    if request.method == 'DELETE':
+        story = Story.query.get(story_id)
+        if story:
+            if story.user_id == current_user.id:
+                db.session.delete(story)
+                db.session.commit()
+                return jsonify({'message': 'Successfully deleted'}), 200
+            else: 
+                return jsonify({'message': 'Users can only delete their own stories'}), 403
+        else:
+            return jsonify({'message': 'Story could not be found'}), 404
